20230807_최종일수도

+++ BusanProject/src/main/java/edu/pnu/repository/BusanFoodRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import edu.pnu.entity.BusanFood;
 
 @Repository
-public interface BusanFoodRepository extends JpaRepository<BusanFood, Long> {
-    Optional<BusanFood> findById(Long id);
+public interface BusanFoodRepository extends JpaRepository<BusanFood, Integer> {
+    Optional<BusanFood> findById(Integer id);
 }
